Add unit tests for connection string handling in AddConnectionString method
